Refactor test utils first part
