+++ roles_royce/applications/bridge_keeper/env.py
@@ -6,7 +6,7 @@ from web3 import Web3
 from web3.types import Address, ChecksumAddress
 
 from roles_royce.applications.utils import check_env_endpoints, custom_config, to_dict
-from tests.utils import fork_unlock_account
+from tests.fork_utils import fork_unlock_account
 
 
 @dataclass
