+++ roles_royce/applications/spark_anti_liquidation_bot/spark_anti_liquidation.py
@@ -27,7 +27,7 @@ env = ENV()
 
 test_mode = env.TEST_MODE
 if test_mode:
-    from tests.utils import top_up_address
+    from tests.fork_utils import top_up_address
 
     w3 = Web3(Web3.HTTPProvider(f"http://localhost:{env.LOCAL_FORK_PORT}"))
     top_up_address(w3, env.BOT_ADDRESS, 1)
