+++ tests/applications/execution_app/test_transaction_builder.py
@@ -3,7 +3,7 @@ import pytest
 from roles_royce.applications.execution_app.transaction_builder import build_transaction
 from roles_royce.applications.execution_app.utils import ENV, ExecConfig
 from tests.fork_fixtures import accounts, local_node_eth
-from tests.utils import assign_role
+from tests.roles_utils import assign_role
 
 dao = "GnosisDAO"
 blockchain = "ETHEREUM"
