+++ tests/development/_test_balancer_vebal.py
@@ -5,7 +5,8 @@ from defi_protocols.constants import ETHEREUM
 from defi_protocols.functions import get_decimals, last_block
 from web3 import HTTPProvider, Web3
 
-from tests.utils import ETH_FORK_NODE_URL, LOCAL_NODE_PORT, fork_reset_state, fork_unlock_account
+from tests.utils import ETH_FORK_NODE_URL, LOCAL_NODE_PORT
+from tests.fork_utils import fork_unlock_account, fork_reset_state
 
 
 def test_balancer_vebal(wallet, block, decimals=True):
