+++ tests/fork_utils.py
@@ -1,0 +1,256 @@
+import atexit
+import codecs
+import json
+import logging
+import shlex
+import shutil
+import socket
+import subprocess
+import time
+import os
+
+from eth_account import Account
+from eth_account.signers.local import LocalAccount
+from web3 import Web3, HTTPProvider
+from web3.exceptions import ContractLogicError
+
+from defabipedia.tokens import erc20_contract
+from tests.simple_safe import SimpleSafe
+
+logger = logging.getLogger(__name__)
+
+
+def gen_test_accounts() -> list[LocalAccount]:
+    # test accounts are generated using the mnemonic:
+    #   "test test test test test test test test test test test junk" and derivation path "m/44'/60'/0'/0"
+    keys = [
+        "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
+        "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d",
+        "0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a",
+        "0x7c852118294e51e653712a81e05800f419141751be58f605c371e15141b007a6",
+        "0x47e179ec197488593b187f80a00eb0da91f1b9d0b13f8733639f19c30a34926a",
+        "0x8b3a350cf5c34c9194ca85829a2df0ec3153be0318b5e2d3348e872092edffba",
+        "0x92db14e403b83dfe3df233f83dfa3a0d7096f21ca9b0d6d6b8d88b2b4ec1564e",
+        "0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1fcdbf7cbf4356",
+        "0xdbda1821b80551c9d65939329250298aa3472ba22feea921c0cf5d620ea67b97",
+        "0x2a871d0798f97d79848a013d4936a73bf4cc922c825d33c1cf7073dff6d409c6",
+    ]
+    return [Account.from_key(key) for key in keys]
+
+
+TEST_ACCOUNTS = gen_test_accounts()
+SCRAPE_ACCOUNT = Account.from_key("0xf214f2b2cd398c806f84e317254e0f0b801d0643303237d97a22a48e01628897")
+
+REMOTE_ETH_NODE_URL = codecs.decode(
+    b"x\x9c\xcb())(\xb6\xd2\xd7O-\xc9\xd0\xcdM\xcc\xcc\xcbK-\xd1K\xd7K\xccI\xceH\xcd\xad\xd4K\xce\xcf\xd5/3\xd2\x0f"
+    b"u)74-6NNu\xb3\xcc\x0f\nH\n\xcb\xccq4\xd4u\xcd3(53+\xf32\n(\x06\x00Q\x92\x17X",
+    "zlib",
+).decode()
+ETH_FORK_NODE_URL = os.environ.get("RR_ETH_FORK_URL", REMOTE_ETH_NODE_URL)
+REMOTE_GC_NODE_URL = "https://rpc.ankr.com/gnosis"
+GC_FORK_NODE_URL = os.environ.get("RR_GC_FORK_URL", REMOTE_GC_NODE_URL)
+ETH_LOCAL_NODE_PORT = 8546
+GC_LOCAL_NODE_PORT = 8547
+ETH_LOCAL_NODE_DEFAULT_BLOCK = 17565000
+GC_LOCAL_NODE_DEFAULT_BLOCK = 30397769
+RUN_LOCAL_NODE = os.environ.get("RR_RUN_LOCAL_NODE", False)
+ETH_LOCAL_NODE_URL = f"http://127.0.0.1:{ETH_LOCAL_NODE_PORT}"
+GC_LOCAL_NODE_URL = f"http://127.0.0.1:{GC_LOCAL_NODE_PORT}"
+
+
+def wait_for_port(port, host="localhost", timeout=5.0):
+    """Wait until a port starts accepting TCP connections."""
+    start_time = time.time()
+    while True:
+        try:
+            s = socket.create_connection((host, port), timeout=timeout)
+            s.close()
+            return
+        except socket.error:
+            time.sleep(0.05)
+            if time.time() - start_time >= timeout:
+                raise socket.error("Timeout waiting for port")
+
+
+class SimpleDaemonRunner(object):
+    def __init__(self, cmd, popen_kwargs=None):
+        self.console = None
+        self.proc = None
+        self.cmd = cmd
+        self.popen_kwargs = popen_kwargs or {}
+
+    def start(self):
+        if self.is_running():
+            raise ValueError("Process is already running")
+        logger.info("Starting daemon: %s %s", self.cmd, self.popen_kwargs)
+        self.proc = subprocess.Popen(shlex.split(self.cmd), **self.popen_kwargs)
+        atexit.register(self.stop)
+
+    def stop(self):
+        if not self.proc:
+            return
+
+        self.proc.terminate()
+        stdout, stderr = self.proc.communicate(timeout=20)
+        retcode = self.proc.returncode
+
+        self.proc = None
+        return retcode
+
+    def is_running(self):
+        return self.proc is not None
+
+
+def fork_unlock_account(w3, address):
+    """Unlock the given address on the forked node."""
+    return w3.provider.make_request("anvil_impersonateAccount", [address])
+
+
+def fork_set_balance(w3, address, amount):
+    return w3.provider.make_request("anvil_setBalance", [address, amount])
+
+
+def fork_reset_state(w3: Web3, url: str, block: int | str = "latest"):
+    """Reset the state of the forked node to the state of the blockchain node at the given block.
+
+    Args:
+        w3: Web3 instance of the local node
+        url: URL of the node from which to fork
+        block: Block number at which to fork the blockchain, or "latest" to use the latest block
+    """
+
+    if isinstance(block, str):
+        if block == "latest":
+            raise ValueError("Can't use 'latest' as fork block")
+    return w3.provider.make_request("anvil_reset", [{"forking": {"jsonRpcUrl": url, "blockNumber": block}}])
+
+
+def run_hardhat():
+    """Run hardhat node in the background."""
+    try:
+        npm = shutil.which("npm")
+        subprocess.check_call([npm, "--version"])
+        if "hardhat" not in json.loads(subprocess.check_output([npm, "list", "--json"])).get("dependencies", {}):
+            raise subprocess.CalledProcessError
+    except subprocess.CalledProcessError:
+        raise RuntimeError("Hardhat is not installed properly. Check the README for instructions.")
+
+    log_filename = "/tmp/rr_hardhat_log.txt"
+    logger.info(f"Writing Hardhat log to {log_filename}")
+    hardhat_log = open(log_filename, "w")
+    npx = shutil.which("npx")
+    node = SimpleDaemonRunner(
+        cmd=f"{npx} hardhat node --show-stack-traces --fork '{ETH_FORK_NODE_URL}' --fork-block-number {ETH_LOCAL_NODE_DEFAULT_BLOCK} --port {ETH_LOCAL_NODE_PORT}",
+        popen_kwargs={"stdout": hardhat_log, "stderr": hardhat_log},
+    )
+    node.start()
+    return node
+
+
+def run_anvil(url, block, port):
+    """Run anvil node in the background"""
+    log_filename = f"/tmp/rr_fork_node_{port}_log.txt"
+    logger.info(f"Writing Anvil log to {log_filename}")
+    log = open(log_filename, "w")
+    node = SimpleDaemonRunner(
+        cmd=f"anvil --accounts 15 -f '{url}' --fork-block-number {block} --port {port}",
+        popen_kwargs={"stdout": log, "stderr": log},
+    )
+
+    node.start()
+    return node
+
+
+class LocalNode:
+    def __init__(self, remote_url: str, port: int, default_block: int):
+        self.remote_url = remote_url
+        self.port = port
+        self.url = f"http://127.0.0.1:{port}"
+        self.default_block = default_block
+        self.w3 = Web3(HTTPProvider(self.url, request_kwargs={"timeout": 30}))
+
+    def reset_state(self):
+        fork_reset_state(self.w3, self.remote_url, self.default_block)
+
+    def unlock_account(self, address: str):
+        fork_unlock_account(self.w3, address)
+
+    def set_block(self, block):
+        """Set the local node to a specific block"""
+        fork_reset_state(self.w3, url=self.remote_url, block=block)
+
+
+def _local_node(request, node: LocalNode):
+    """Run a local node_daemon for testing"""
+    if RUN_LOCAL_NODE:
+        node_daemon = run_anvil(node.remote_url, node.default_block, node.port)
+
+        def stop():
+            node_daemon.stop()
+
+        request.addfinalizer(stop)
+
+    wait_for_port(node.port, timeout=20)
+
+    class LatencyMeasurerMiddleware:
+        def __init__(self, make_request, w3):
+            self.w3 = w3
+            self.make_request = make_request
+
+        def __call__(self, method, params):
+            import time
+
+            start_time = time.monotonic()
+            response = self.make_request(method, params)
+            logger.debug("Web3 time spent in %s: %f seconds", method, time.monotonic() - start_time)
+            return response
+
+    node.w3.middleware_onion.add(LatencyMeasurerMiddleware, "latency_middleware")
+    node.reset_state()
+    return node
+
+
+def steal_token(w3, token, holder, to, amount):
+    """Steal tokens from a holder sending them to another address (sends 1 ETH to the holder first)"""
+    fork_unlock_account(w3, holder)
+    top_up_address(w3=w3, address=holder, amount=1)
+    ctract = erc20_contract(w3, token)
+    tx = ctract.functions.transfer(to, amount).transact({"from": holder})
+    return tx
+
+
+def create_simple_safe(w3: Web3, owner: LocalAccount) -> SimpleSafe:
+    """Create a Safe with one owner and 1 ETH in balance"""
+
+    safe = SimpleSafe.build(owner, w3)
+    top_up_address(w3=w3, address=safe.address, amount=1)
+    fork_unlock_account(w3, safe.address)
+    return safe
+
+
+def top_up_address(w3: Web3, address: str, amount: int) -> None:
+    """Top up an address with ETH"""
+    if amount > (w3.eth.get_balance(SCRAPE_ACCOUNT.address) * 1e18) * 0.99:
+        raise ValueError("Not enough ETH in the faucet account")
+    try:
+        w3.eth.send_transaction({"to": address, "value": Web3.to_wei(amount, "ether"), "from": SCRAPE_ACCOUNT.address})
+    except ContractLogicError:
+        raise Exception("Address is a smart contract address with no payable function.")
+
+
+def steal_safe(w3, safe_address, new_owner_address):
+    """
+    Return a SimpleSafe instance from an existing safe with a new owner.
+
+    The new owner will be added to the owners list, and the threshold will be set to 1,
+    so it will be able to sign and execute transactions.
+    """
+
+    safe = SimpleSafe(safe_address, w3=w3, signer_key=new_owner_address)
+
+    fork_unlock_account(w3, safe_address)
+    fork_set_balance(w3, safe.address, 10000000000000000000)  # give some gas: 10 ETH
+
+    # set thresshold to 1 and add owner
+    safe.contract.functions.addOwnerWithThreshold(new_owner_address, 1).transact({"from": safe.address})
+    return safe
