+++ tests/protocols/cowswap/test_contract_methods.py
@@ -2,8 +2,8 @@ from defabipedia.types import Chain
 from roles_royce import roles
 from roles_royce.protocols.cowswap.contract_methods import SignOrder
 from roles_royce.protocols.cowswap.utils import create_order_api, quote_order_api
-from tests.roles import apply_presets, deploy_roles, setup_common_roles
-from tests.utils import create_simple_safe
+from tests.roles_utils import apply_presets, deploy_roles, setup_common_roles
+from tests.fork_utils import create_simple_safe
 from tests.fork_fixtures import accounts
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth
