+++ tests/protocols/cowswap/test_methods_swap.py
@@ -3,8 +3,8 @@ from defabipedia.types import Chain
 from roles_royce import roles
 from roles_royce.protocols.cowswap.methods_swap import create_order_and_swap, swap
 from roles_royce.protocols.cowswap.utils import SwapKind, Order, quote_order_api
-from tests.roles import apply_presets, deploy_roles, setup_common_roles
-from tests.utils import create_simple_safe
+from tests.roles_utils import apply_presets, deploy_roles, setup_common_roles
+from tests.fork_utils import create_simple_safe
 from tests.fork_fixtures import accounts
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth
