+++ tests/roles_utils.py
@@ -1,5 +1,7 @@
 import json
 
+from web3.types import TxReceipt
+
 from defabipedia.types import Chain
 from web3 import Web3
 
@@ -7,7 +9,9 @@ from roles_royce.evm_utils import roles_abi, roles_bytecode
 from roles_royce.generic_method import TxData
 from roles_royce.utils import MULTISENDS, to_checksum_address
 
-from .utils import TEST_ACCOUNTS, SimpleSafe
+from .simple_safe import SimpleSafe
+from .fork_utils import TEST_ACCOUNTS, top_up_address
+from .utils import to_hex_32_bytes
 
 
 def deploy_roles(w3: Web3, avatar):
@@ -66,3 +70,33 @@ def apply_presets(safe: SimpleSafe, roles_ctract, json_data, replaces=None):
         for replacer in replaces:
             data = data.replace(replacer[0], replacer[1])
         safe.send(txs=[TxData(contract_address=roles_ctract.address, data=data)])
+
+
+def assign_role(local_node, avatar_safe_address: str, roles_mod_address: str, role: int, asignee: str) -> TxReceipt:
+    asignee_32_bytes = to_hex_32_bytes(asignee)
+    role_32_byes = to_hex_32_bytes(role)
+    a = asignee_32_bytes[2:]
+    calldata_to_assign_role = (
+        f"0xa6edf38f"
+        f"{asignee_32_bytes[2:]}"
+        f"0000000000000000000000000000000000000000000000000000000000000060"
+        f"00000000000000000000000000000000000000000000000000000000000000a0"
+        f"0000000000000000000000000000000000000000000000000000000000000001"
+        f"{role_32_byes[2:]}"
+        f"0000000000000000000000000000000000000000000000000000000000000001"
+        f"0000000000000000000000000000000000000000000000000000000000000001"
+    )
+
+    tx_to_assign_role = {
+        "from": avatar_safe_address,
+        "to": roles_mod_address,
+        "data": calldata_to_assign_role,
+        "value": "0",
+    }
+
+    local_node.unlock_account(avatar_safe_address)
+    # The amount of ETH of the Avatar address is increased
+    top_up_address(local_node.w3, address=avatar_safe_address, amount=1)
+    tx_hash = local_node.w3.eth.send_transaction(tx_to_assign_role)
+    tx_receipt = local_node.w3.eth.wait_for_transaction_receipt(tx_hash, timeout=5)
+    return tx_receipt
