+++ tests/simple_safe.py
