+++ tests/test_spark.py
@@ -6,7 +6,8 @@ from roles_royce.constants import ETHAddr
 from roles_royce.protocols.eth import spark
 from roles_royce.toolshed.protocol_utils.spark.utils import SparkUtils
 
-from .utils import create_simple_safe, get_balance, steal_token
+from .utils import get_balance
+from .fork_utils import steal_token, create_simple_safe
 from .fork_fixtures import accounts
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth
