+++ tests/toolshed/disassembling/test_disassembling_aura.py
@@ -5,7 +5,8 @@ from defabipedia.types import Chain
 
 from roles_royce.constants import ETHAddr
 from roles_royce.toolshed.disassembling import AuraDisassembler
-from tests.utils import get_balance, top_up_address
+from tests.utils import get_balance
+from tests.fork_utils import top_up_address
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth
