+++ tests/toolshed/disassembling/test_disassembling_balancer.py
@@ -8,8 +8,8 @@ from roles_royce.constants import ETHAddr
 from roles_royce.evm_utils import erc20_abi
 from roles_royce.roles_modifier import GasStrategies, set_gas_strategy
 from roles_royce.toolshed.disassembling import BalancerDisassembler
-from tests.roles import apply_presets, deploy_roles, setup_common_roles
-from tests.utils import create_simple_safe, steal_token, top_up_address
+from tests.roles_utils import apply_presets, deploy_roles, setup_common_roles
+from tests.fork_utils import steal_token, create_simple_safe, top_up_address
 from tests.fork_fixtures import accounts
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth
