+++ tests/toolshed/disassembling/test_disassembling_spark.py
@@ -10,8 +10,9 @@ from roles_royce.roles_modifier import GasStrategies, set_gas_strategy
 from defabipedia.types import Chain
 from defabipedia.spark import ContractSpecs
 
-from tests.roles import apply_presets, deploy_roles, setup_common_roles
-from tests.utils import create_simple_safe, get_balance, steal_token, top_up_address
+from tests.roles_utils import apply_presets, deploy_roles, setup_common_roles
+from tests.utils import get_balance
+from tests.fork_utils import steal_token, create_simple_safe, top_up_address
 from tests.fork_fixtures import accounts
 from tests.fork_fixtures import local_node_eth_replay as local_node_eth, local_node_gc
